+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - State.Builder added.
 ### is.codion.framework.domain
 - OrderBy.Builder, NullOrder now parameter.
+- OrderBy.OrderByColumn.ignoreCase() added.
 ### is.codion.swing.common.ui
 - InputDialogBuilder.inputValid() and inputValidator() renamed valid() and validator() respectively.
 ### is.codion.swing.fraework.ui